refactor: AppBottomBar
